Add GitHub personal access token as CLI argument
We're requiring a GitHub personal access token as an additional CLI
argument as the Git tooling helper that we're using now requires it
to be passed as an argument when creating a new instance of `Git`.
This is so that it can clone and push to private repositories on
GitHub.