Update README to reflect new CLI options
The new Git branch name and GitHub personal access token options
have been documented.

Now that settings are explicitly passed to the CLI, the concept of
"transformations" is redundant when it comes to this tool. Aside
from the name of the tool itself (which needs to be changed), all
mentions of "transformations" have been removed from this README.