Implement yargs for options
Currently, options have to be passed into the CLI tool in a particular
order. The options, based on how they were passed in, are assigned to
a variable and used throughout the script. By using yargs, we can avoid
all of that complexity and users can pass in the option with a flag in
whatever order they please.