add @quarterto/lint settings
this code style is going to be divisive! but, the general setup is one i've found to be productive.

since we have no currently-accepted ft.com-wide style preferences, go with this for now?

i think the main point of contention is going to be semicolons, which, with linting and formatting has very little chance of running into the bad cases