Retry requests on abuse limits as well
We've been running into abuse limiting when creating many PRs for a
workspace, presumably as a result of secondary rate limiting. From
GitHub's best practises documentation:

> Requests that create content which triggers notifications, such as
issues, comments and pull requests, may be further limited and will not
include a Retry-After header in the response. Please create this content
at a reasonable pace to avoid further limiting.

Requests are retried at a minimum of 5 seconds after the original, which
is probably what we want even if the documentation is (intentionally)
vague.

Closes #149